feat(#1023): ICA: ExecFn type duplicated with divergent contracts
