+++ .pi/extensions/supervisor/checks/dead-code.ts
@@ -6,8 +6,14 @@
 // If knip is unavailable, gracefully degrades with status "no_knip".
 // The auditor agent then uses ripgrep_search / structural_search as fallback.
 
+/** Exec function type for subprocess calls (3-field return — code, stdout, stderr) */
+export type ExecFn = (
+	cmd: string,
+	args: string[],
+	opts?: Record<string, unknown>,
+) => Promise<{ code: number; stdout: string; stderr: string }>;
+
 import {
-	type ExecFn,
 	getChangedFilesFromGitDiff,
 	filterItemsToChangedFiles,
 	sumLines,
