+++ .pi/extensions/supervisor/checks/package-safety.ts
@@ -1,6 +1,12 @@
 import { readFile } from "node:fs/promises";
 import { join } from "node:path";
-import type { ExecFn } from "./shared.ts";
+
+/** Exec function type for subprocess calls (3-field return — code, stdout, stderr) */
+type ExecFn = (
+	cmd: string,
+	args: string[],
+	opts?: Record<string, unknown>,
+) => Promise<{ code: number; stdout: string; stderr: string }>;
 
 // ─── Package Safety Check ──────────────────────────────────────────
 // Deterministic package age validation for npm install safety.
