+++ .pi/extensions/supervisor/checks/requirements-traceability.ts
@@ -6,7 +6,13 @@
 
 import { existsSync } from "node:fs";
 import { join, extname, basename, dirname } from "node:path";
-import type { ExecFn } from "./shared.ts";
+
+/** Exec function type for subprocess calls (3-field return — code, stdout, stderr) */
+export type ExecFn = (
+	cmd: string,
+	args: string[],
+	opts?: Record<string, unknown>,
+) => Promise<{ code: number; stdout: string; stderr: string }>;
 
 // ─── Types ──────────────────────────────────────────────────────────
