+++ .pi/extensions/supervisor/test/duplicate-code.test.mts
@@ -20,7 +20,7 @@ import {
 	buildResult,
 	runDuplicateCheck,
 } from "../checks/duplicate-code.ts";
-import type { ExecFn } from "../checks/shared.ts";
+import type { ExecFn } from "../checks/duplicate-code.ts";
 
 // ═══════════════════════════════════════════════════════════════════════
 // Helper: build a JscpdClone for use in tests
